+++ nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=ecf11dbe
+build.xml.script.CRC32=7f3171db
+build.xml.stylesheet.CRC32=8064a381@1.80.1.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=ecf11dbe
+nbproject/build-impl.xml.script.CRC32=2198ddb6
+nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
