+++ tech/kunonu/digitalgate/awali/package-info.java
@@ -1,4 +1,4 @@
-package tech.kunonu.digitalgate.awali;
+package kunonu.digitalgate.awali;
 
 class about
 {
